[STEP3] CardRank, CardSuit 파일 제거 (Card 파일 내로 위치 변경)
