--- src/main/kotlin/blakjack/domain/CardSuit.kt
@@ -1,5 +1,0 @@
-package blakjack.domain
-
-enum class CardSuit {
-    HEART, DIAMOND, SPADE, CLOVER
-}
